+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ### Added
 * GFX index (under feature flag).
+* Updated Korean translation (Contributor: [gyhs(NIKA)](https://github.com/gyhs)).
 
 ### Fixed
 * NSB tank tech tree (NSB_armor.txt) doesn't preview (#71).